Bugfix for MINP: it is a long
When implementing the MINP (MSTA input) it was overseen,
that MSTA is defined as 32 bit (where only 15 bits are used).
Correct this: use long instead of short.

+++ motorApp/SoftMotorSrc/devSoft.cc
@@ -254,7 +254,7 @@ void soft_dinp_func(struct motorRecord *mr, short newdinp)
     }
 }
 
-void soft_minp_func(struct motorRecord *mr, short newminp)
+void soft_minp_func(struct motorRecord *mr, long newminp)
 {
     msta_field status;
     unsigned int old_RA_DONE;
